28 updates to local rootfs package files are not detected (#31)
* Detect changes in local debian packages

* Avoid rebuilding when there are no local debian packages

* Check Last-Modified field of packages online to see whether they need to be updated

+++ src/abstract_builders/debian_rootfs_builder.py
@@ -320,13 +320,55 @@ class Debian_RootFS_Builder(File_System_Builder):
             )
             return
 
+        # Create a list of local packages and a list of last modified timestamps of online packages
+        local_pkgs = []
+        online_pkg_timestamps = []
+        for uri in self.block_cfg.project.addl_ext_pkgs:
+            if urllib.parse.urlparse(uri).scheme == "file":
+                # This package is provided locally
+                local_pkg_path = pathlib.Path(urllib.parse.urlparse(uri).path)
+                # Append file to list of local packages
+                local_pkgs.append(local_pkg_path)
+            elif urllib.parse.urlparse(uri).scheme in ["http", "https"]:
+                # This package is provided online
+                try:
+                    online_pkg_timestamps.append(File_Downloader.get_last_modified(url=uri))
+                except RuntimeError:
+                    pretty_print.print_warning(
+                        "Updates to the following package cannot be detected because no 'Last-Modified' field "
+                        "could be retrieved from the header. Use a different server or trigger rebuilds "
+                        f"manually: {uri}"
+                    )
+
+        # Check whether a rebuild is necessary due to updated local packages
+        rebuild_bc_local_pkgs = False
+        if local_pkgs:
+            rebuild_bc_local_pkgs = Build_Validator.check_rebuild_bc_timestamp(
+                src_search_list=local_pkgs,
+                out_timestamp=self._build_log.get_logged_timestamp(
+                    identifier=f"function-{inspect.currentframe().f_code.co_name}-success"
+                ),
+            )
+
+        # Check whether a rebuild is necessary due to updated online packages
+        rebuild_bc_online_pkgs = False
+        if online_pkg_timestamps:
+            rebuild_bc_online_pkgs = max(online_pkg_timestamps) > self._build_log.get_logged_timestamp(
+                identifier=f"function-{inspect.currentframe().f_code.co_name}-success"
+            )
+
         # Check whether the extra packages need to be added
         packages_already_added = (
             self._build_log.get_logged_timestamp(identifier=f"function-{inspect.currentframe().f_code.co_name}-success")
             != 0.0
         )
-        if packages_already_added and not self._build_validator.check_rebuild_bc_config(
-            keys=[["blocks", self.block_id, "project", "addl_ext_pkgs"]]
+        if (
+            packages_already_added
+            and not rebuild_bc_local_pkgs
+            and not rebuild_bc_online_pkgs
+            and not self._build_validator.check_rebuild_bc_config(
+                keys=[["blocks", self.block_id, "project", "addl_ext_pkgs"]]
+            )
         ):
             pretty_print.print_build(
                 "No need to install additional packages from external *.deb files. No altered source files detected..."
